Rename executor to task
