+++ recce/tasks/__init__.py
@@ -1,0 +1,4 @@
+from .core import Task
+from .profile import ProfileDiffTask
+from .query import QueryTask, QueryDiffTask
+from .valuediff import ValueDiffTask
